Fix conflicts & update to match latest code
