+++ test/fixtures/outputs/generateStops-output.json
@@ -28,8 +28,8 @@
       ["R1720058", "#bdd7e7"],
       ["R1720059", "#bdd7e7"],
       ["R1720063", "#bdd7e7"],
-      ["R1720066", "#bdd7e7"],
       ["R1720065", "#bdd7e7"],
+      ["R1720066", "#bdd7e7"],
       ["R1720055", "#bdd7e7"],
       ["R1720051", "#bdd7e7"],
       ["R1720068", "#bdd7e7"],
@@ -41,23 +41,23 @@
       ["R1720054", "#6baed6"]
     ],
     [
-      ["R1720060", "#eff3ff"],
-      ["R1720052", "#eff3ff"],
       ["R1720056", "#eff3ff"],
       ["R1720058", "#eff3ff"],
-      ["R1720061", "#eff3ff"],
       ["R1720062", "#eff3ff"],
+      ["R1720061", "#eff3ff"],
+      ["R1720052", "#eff3ff"],
+      ["R1720060", "#eff3ff"],
       ["R1720053", "#eff3ff"],
       ["R1720065", "#eff3ff"],
       ["R1720063", "#eff3ff"],
-      ["R1720066", "#eff3ff"],
       ["R1720064", "#eff3ff"],
+      ["R1720066", "#eff3ff"],
       ["R1720055", "#eff3ff"],
       ["R1720059", "#bdd7e7"],
       ["R1720067", "#bdd7e7"],
       ["R1720057", "#bdd7e7"],
-      ["R1720068", "#bdd7e7"],
       ["R1720051", "#bdd7e7"],
+      ["R1720068", "#bdd7e7"],
       ["R1720054", "#6baed6"]
     ],
     [
@@ -110,8 +110,8 @@
       ["R1720058", 6],
       ["R1720059", 6],
       ["R1720063", 6],
-      ["R1720066", 6],
       ["R1720065", 6],
+      ["R1720066", 6],
       ["R1720055", 6],
       ["R1720051", 6],
       ["R1720068", 6],
@@ -123,23 +123,23 @@
       ["R1720054", 9]
     ],
     [
-      ["R1720060", 3],
-      ["R1720052", 3],
       ["R1720056", 3],
       ["R1720058", 3],
-      ["R1720061", 3],
       ["R1720062", 3],
+      ["R1720061", 3],
+      ["R1720052", 3],
+      ["R1720060", 3],
       ["R1720053", 3],
       ["R1720065", 3],
       ["R1720063", 3],
-      ["R1720066", 3],
       ["R1720064", 3],
+      ["R1720066", 3],
       ["R1720055", 3],
       ["R1720059", 6],
       ["R1720067", 6],
       ["R1720057", 6],
-      ["R1720068", 6],
       ["R1720051", 6],
+      ["R1720068", 6],
       ["R1720054", 9]
     ],
     [
@@ -195,8 +195,8 @@
       ["R1720058", 1],
       ["R1720059", 1],
       ["R1720063", 1],
-      ["R1720066", 1],
       ["R1720065", 1],
+      ["R1720066", 1],
       ["R1720055", 1],
       ["R1720051", 1],
       ["R1720068", 1],
@@ -208,23 +208,23 @@
       ["R1720054", 1]
     ],
     [
-      ["R1720060", 1],
-      ["R1720052", 1],
       ["R1720056", 1],
       ["R1720058", 1],
-      ["R1720061", 1],
       ["R1720062", 1],
+      ["R1720061", 1],
+      ["R1720052", 1],
+      ["R1720060", 1],
       ["R1720053", 1],
       ["R1720065", 1],
       ["R1720063", 1],
-      ["R1720066", 1],
       ["R1720064", 1],
+      ["R1720066", 1],
       ["R1720055", 1],
       ["R1720059", 1],
       ["R1720067", 1],
       ["R1720057", 1],
-      ["R1720068", 1],
       ["R1720051", 1],
+      ["R1720068", 1],
       ["R1720054", 1]
     ],
     [
